cpu: Make sure interrupts do not disappear
When the guest is processing an interrupt, an intercept can happen which
will stop the injection and set EXITINTINFO.V = 1.

That then means that the interrupt needs to be reinjected. Usually, when
the hypervisor sees the valid bit, it will re-inject the interrupt.

However, with Alternate Injection, the hypervisor can not re-inject the
interrupt into the guest directly but the SVSM needs to do that.

At the same time, the busy bit gets set so clear it too.

Signed-off-by: Melody Wang <huibo.wang@amd.com>